Added DomainUnloaded event.
When using the NUnit3 test adapter, the process doesn't exit, the app domain is simply unloaded. This means that the  ProcessExit event is never raised. The DomainUnload event is called always. To make sure we don't add double hits, we clear the _markers collection after writing (in case the event handler is triggered twice).