Add support for an intermediate result.
This commits adds the option `IntermediateResult`, which will write the
raw coverage result to an intermediate file, and merge the result of the
next run with that file. This is useful in multi-project solutions.

Eventually, the last coverage run will produce a report with the
combined results of all the runs.