Updating coverlet collector GetPropertyValueWrapper method to add a fallback to get the test source property when the GetPropertyValue<T> method is not found.
If the GetProperties method is also not found, a missingMethodException will also be raised.
If the GetProperties method doesn't return the property, we rethrow the exception.

I've tested both methods return the TestSources property when run using .NET Core SDK 2.2.400.

This is a speculative fix to see if GetProperties method exists for .NET Core SDK 2.1.400.