Don't call GetAwaiter().GetResult() on a ValueTask
Modified calls to GetAwaiter().GetResult() on a ValueTask<int> to use
"await" instead.

As it turns out, this is more of a stylistic change than anything else,
as we're not actually dependent on the result of the calls to the
various AsyncAwaitValueTask methods, since we're only interested in
assessing coverage, but it's a good change nonetheless.