Upgrade LINQ and EF dependencies
After attempting to use IAsyncEnumerable<T> in the coverlet.core.tests
project, I was met with the following error message:

The type 'IAsyncEnumerable<T>' exists in both 'System.Interactive.Async,
Version=3.0.3000.0, Culture=neutral, PublicKeyToken=94bc3704cddfc263'
and 'System.Runtime, Version=5.0.0.0, Culture=neutral,
PublicKeyToken=b03f5f7f11d50a3a'

After some research, I determined that the root cause was the
dependency on an old version of LinqKit.Microsoft.EntityFrameworkCore,
which, in turn, includes a dependency on an old version of Entity
Framework.

To solve that problem, I've upgraded the dependency on
LinqKit.Microsoft.EntityFrameworkCore to the latest version in NuGet
(5.0.23), as well as the dependency on Microsoft.Extensions.Logging.Abstractions,
since it didn't appear to be possible to upgrade one without the other.
Since it looks like Coverlet, generally, has moved on to .NET 5, this
appears to be non-problematic.

Afterward, all tests yielded the same result, but I was then able to
begin using IAsyncEnumerable<T> normally.