Fixing link for Deterministic build docs had an extra "Documentation" in the path
just above my commit message the following text appears...

>  ProTip! Great commit summaries contain fewer than 50 characters. Place extra information in the extended description.

Great commit summaries explain what the commit was for in plain language, without abbreviations and other shortcuts, and should be at _least_ 50 characters to avoid frustratingly terse and useless commit messages like "changed path" or "widget update". Really whomever added that Protip is not a boob and not a pro.